==> 2423 Einspeisevergütung/1_marktpreis_bfe.ipynb <==
import pandas as pd
import altair as alt

start_date = '2017-01-01'
end_date = '2023-12-31'
years = 7

def to_chf_kwh(df, column='Price_CHF_MWh'):
    return df[column] / 1000 * 100

def to_date(df):
    return pd.PeriodIndex(df.Year.astype(str) + df.Period, freq="Q").to_timestamp()


Referenzmarktpreis 2015–17, gewichtet nach ganzem Produktionsvolumen. Nicht speziell nach PV-Produktion.

middle = pd.read_csv("daten/ogd42_rmp_quartalspreise_2015-2017.csv").rename({'Jahr': 'Year'}, axis=1)
middle['chf_kwh'] = to_chf_kwh(middle) 
middle['Datum'] = to_date(middle) 
middle.head()

Referezmarktpreis 2018–23, gewichtet nach PV-Produktionsvolumen:

new = pd.read_csv("daten/ogd60_rmp_quartalspreise.csv")
new['chf_kwh'] = to_chf_kwh(new, column='Price_pv_CHF_MWh')
new['Datum'] = to_date(new)
new.head()

newest = pd.read_csv("daten/newest.csv")
newest['chf_kwh'] = to_chf_kwh(newest)
newest['Datum'] = to_date(newest)
newest

d1 = pd.concat([middle, new, newest])[['Datum', 'Year', 'Period', 'chf_kwh']]
d1

# PeriodIndex to string
d2 = d1.copy()
d2.loc[d2.Period == 'Q1', 'weight'] = 0.135 / years
d2.loc[d2.Period == 'Q2', 'weight'] = 0.365 / years
d2.loc[d2.Period == 'Q3', 'weight'] = 0.365 / years
d2.loc[d2.Period == 'Q4', 'weight'] = 0.135 / years
d2

## Mindestpreis == 0 Rappen
d2['chf_kwh_0'] = d2['chf_kwh']
d2.loc[d2.chf_kwh < 0, 'chf_kwh_0'] = 0

## Mindestpreis == 4.6 Rappen
d2['chf_kwh_4_6'] = d2['chf_kwh']
d2.loc[d2.chf_kwh < 4.6, 'chf_kwh_4_6'] = 4.6

## Mindestpreis == 8 Rappen
d2['chf_kwh_8'] = d2['chf_kwh']
d2.loc[d2.chf_kwh < 8, 'chf_kwh_8'] = 8


d3 = d2[['Datum', 'chf_kwh_0', 'chf_kwh_4_6', 'chf_kwh_8', 'weight']]
d3.to_csv("calc/marktpreis_quarter.csv", index=False)
d3.head()

alt.Chart(d3.melt(id_vars=['Datum', 'weight']).rename({'value': 'chf_kwh', 'variable': 'mindestpreis'}, axis=1).query('Datum >= "'+start_date+'"')).mark_line(interpolate='step-after').encode(
    alt.X('Datum').axis(format='%Y-%m'),
    alt.Y('chf_kwh', title='CHF/kWh'),
    alt.Color('mindestpreis', title='Mindestpreis in Rappen'),
).properties(width=800, height=400)

### 7-Jahres-Mittel nach gewichtetem Mindestpreis

d4 = d3.query('Datum >= "'+start_date+'" & Datum <= "'+end_date+'"')
mittel_0 = (d4.chf_kwh_0 * d4.weight).sum()
mittel_0

mittel_4_6 = (d4.chf_kwh_4_6 * d4.weight).sum()
mittel_4_6

mittel_8 = (d4.chf_kwh_8 * d4.weight).sum()
mittel_8

%store mittel_0
%store mittel_4_6
%store mittel_8
