==> 2423 Einspeisevergütung/2_effektiver_preis.ipynb <==
import pandas as pd
import geopandas as gpd
import altair as alt

start = '2017-01-01'
end = '2023-12-31'

## Mittlerer Strompreis der letzen sieben Jahre

* Jahre 2017 bis 2023
* Pro Gemeinde
* Wenn für ein oder mehrere Jahre kein Preis verfügbar ist, nehmen wir den Schnitt der verbleibenden Jahre

## Mittlerer Marktpreis aus 1_marktpreis_bfe.ipynb
%store -r mittel_0
%store -r mittel_4_6
%store -r mittel_8
marktpreis =  mittel_4_6
marktpreis

Daten der Elektrizitätswerke der letzten 7 Jahre von [pvtarif.ch](https://www.vese.ch/pvtarif/)

Als Vergleichspreis nehmen wir `energy1` für die Jahre 2017 bis 2023 sowie die Vergütung für den herkunftsnachweis.

ew = pd.read_csv('daten/ew.csv')
ew['strompreis'] = ew['energyAuto1'] + ew['ecoAuto1']
ew['Datum'] = pd.to_datetime(ew['year'], format='%y')

ew = ew[['Datum', 'nrElcom', 'strompreis', 'nomEw', 'year']]
ew = ew.query('Datum >= "'+start+'" & Datum <= "'+end+'"')
ew.query('strompreis.isna()')

Pro Elektrizitätswerk den mittleren Preis der Jahre 2017 bis 2023 berechnen:

ew1 = ew.groupby('nrElcom').agg({
    'strompreis': 'mean', ## Mittelwert über sieben Jahre
    'nomEw': 'first',
}, skipna=True).reset_index()
ew1.query('strompreis.isna()')

Elektrizitäswerke nach Gemeinden von [pvtarif.ch](https://www.vese.ch/pvtarif/):

gd = pd.read_csv('daten/gd.csv')
gd = gd[['nrElcom', 'idofs', 'Gemeinde']].rename({'idofs': 'ID'}, axis=1)
gd

Für jedes Jahr, Elektrizitätswerk und jede Gemeinde eine eigene Zeile. Einzelne Gemeinden haben mehrere Anbieter. Manche Anbieter haben nicht für jedes Jahr Daten.

merged = pd.merge(ew, gd, on='nrElcom', how='outer')
merged.sort_values(['Gemeinde', 'nomEw', 'year']).to_excel('calc/ew_gd.xlsx', index=False)
merged.query('strompreis.isna()')

Wir gehen davon aus, dass der Strom in jeder Gemeinde an den Anbieter verkauft wird, der den höchsten Preis bietet:

d1 = merged.groupby(['ID', 'year']).agg({
    'strompreis': 'max', # Höchster Preis pro Gemeinde und Jahr
    'Gemeinde': 'first'
}, skipna=True).reset_index()
d1.sort_values(['Gemeinde', 'year']).to_excel('calc/ew_gd_2.xlsx', index=False)
d1.query('Gemeinde == "Davos"')

einwohner = pd.read_csv("daten/einwohner.csv")
einwohner.head()

d1a = pd.merge(d1, einwohner[['ID', 'Einwohner']], on='ID', how='left')
d1a.to_csv('calc/preis_gemeinde_jahr.csv', index=False)
d1a

Mittlerer Preis pro Gemeinde über die Jahre 2017 bis 2023:

d2 = d1a.groupby('ID').agg({
    'strompreis': 'mean', # Mittelwert über sieben Jahre
    'Gemeinde': 'first',
    'Einwohner': 'first'
}, skipna=True).reset_index() # Fehlen Jahre, wird der Mittelwert der restlichen Jahre berechnet
d2.query('Gemeinde == "Davos"')

Nach Einwohner gewichteter Mittelwert.   
Einwohnerzahlen von https://www.bfs.admin.ch/bfs/de/home/statistiken/regionalstatistik/regionale-portraets-kennzahlen/gemeinden.html

d3 = d2.copy()
d3.to_csv('calc/preis_gemeinde.csv', index=False)
d3

d3['weight'] = d3['Einwohner'] / d3['Einwohner'].sum()
d3

(d3.strompreis * d3.weight).sum()
