==> 2423 Einspeisevergütung/4_preiskurven.ipynb <==
import pandas as pd
import altair as alt

start_date= '2017-01-01'

## Vergleich Marktpreis und effektive Vergütung über die Zeit

Wann lagen die Vergütungen über dem Marktpreis und wann darunter?

d1 = pd.read_csv("calc/marktpreis_quarter.csv").query("Datum >= @start_date")
d1 = d1[['Datum', 'chf_kwh_0', 'chf_kwh_4_6', 'chf_kwh_8']]
d1['year'] = pd.to_datetime(d1['Datum']).dt.year
d1

Marktpreis:

markt = alt.Chart(d1).mark_line(interpolate='step-after').encode(
    alt.X('Datum:T'),
    alt.Y('chf_kwh_4_6:Q'),
    color=alt.value("#ae49a2")
)
markt

Nach Einwohner gewichteter, mittlere effektive Vergütung pro Jahr:

d3 = pd.read_csv("calc/preis_gemeinde_jahr.csv")
d3['Datum'] = pd.to_datetime(d3.year, format='%y')
d3['year'] = pd.to_datetime(d3.Datum).dt.year
d3

# Gewichteter Mittelwert nach Jahr. Gewichtung nach Anzahl Einwohner
total_einwohner = d3.groupby('ID').Einwohner.first().sum()
total_einwohner



d4 = d3.copy()
d4['weight'] = d4.Einwohner / total_einwohner
d4 = d4.groupby('year').apply(lambda x: (x['strompreis'] * x['weight']).sum()).reset_index().rename({0: 'effektive'}, axis=1)
d4

diff = pd.merge(d1, d4[['year', 'effektive']], on='year', how='left')
diff


diff = diff.query('Datum >= "'+start_date+'"')
diff.loc[diff.chf_kwh_4_6 >= diff.effektive, 'winner'] = 'Marktpreis'
diff.loc[diff.chf_kwh_4_6 >= diff.effektive, 'lower'] = diff.effektive
diff.loc[diff.chf_kwh_4_6 >= diff.effektive, 'upper'] = diff.chf_kwh_4_6
diff.loc[diff.chf_kwh_4_6 < diff.effektive, 'winner'] = 'Versorger'
diff.loc[diff.chf_kwh_4_6 < diff.effektive, 'lower'] = diff.chf_kwh_4_6
diff.loc[diff.chf_kwh_4_6 < diff.effektive, 'upper'] = diff.effektive

diff['Date_start'] = diff['Datum']
diff['Date_end'] = diff['Datum'].shift(-1)

diff.to_excel('diff.xlsx', index=False)
diff

diff_chart = alt.Chart(diff).mark_rect(clip=True).encode(
    alt.Y('lower:Q'),
    alt.Y2('upper:Q'),
    alt.X('Date_start:T').scale(domain=["2017-01-01", "2023-12-31"]),
    alt.X2('Date_end:T'),
    alt.Color('winner:N').scale(range=['#d485c8', '#39ac9f']).legend(None),
)
diff_chart

ews = alt.Chart(diff).mark_line(interpolate='step-after').encode(
    alt.X('Datum:T').title(None).axis(grid=False, format="%y", values=['2017-01-01', '2018-01-01', '2019-01-01', '2020-01-01', '2021-01-01', '2022-01-01', '2023-01-01', '2024-01-01']),
    alt.Y('effektive:Q').title(None).axis(values=[0,5,10,15,20,25,30,35,40]),
    color=alt.value("#006d64")
)
ews

def chart(width):
    return (diff_chart + markt + ews).properties(width=width, height=300+width*0.2, autosize={'type': 'fit', 'contains': 'padding'}).configure_axis(
    labelFont="GT America",
    labelFontSize=12,
    titleFont="GT America",
    titleFontSize=14,
    titleFontWeight=400,
    gridColor="#d4d6dd",
    labelColor="#6e6e7e",
    ticks=False,
    domain=False,
    labelPadding=5
).configure_text(
    fontSize=14,
    font="GT America"
).configure_axisX(
    ticks=True
).configure_view(strokeWidth=0
)
chart(370).save('mw.svg')
chart(370)

chart(630).save('cw.png', scale_factor=4)
chart(630)

Durchschnittliche effektive Vergütung während der Energiekrise (Q4-2021 bis Q4-2022):

d4

(d4.query('year == 2022').effektive.iloc[0] * 4 + d4.query('year == 2021').effektive.iloc[0])/5
